getline my old friend. Never Forget About While Statements
We’re taking a whack at getline. Opening a string input stream for the
string iArray. Adding a while loop instead of a for loop. A while loop
is better for this function because we set the while condition to check
if the input stream is still open. The while loop’s condition will
catch when the user has stopped inputing and pressed enter.